requirements.txt
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.1
requests==2.31.0
beautifulsoup4==4.12.2
openpyxl==3.1.2
apscheduler==3.10.4
selenium==4.15.2
undetected-chromedriver==3.5.3